Move depot download stuff to separate class
