+++ src/aoe2de_patcher.py
@@ -1,8 +1,8 @@
 from app import App
 
 if __name__ == '__main__':
-  version_major = 2
-  version_minor = 12
+    version_major = 2
+    version_minor = 12
 
-  app = App(version_major, version_minor)
-  app.start()
+    app = App(version_major, version_minor)
+    app.start()
